+++ src/test/java/io/vultr/cld/paycons/persistence/repo/TxRepoTest.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.test.context.ActiveProfiles;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.UUID;
@@ -29,14 +30,18 @@ class TxRepoTest {
     @Test
     void givenTxRepo_whenSavingNewEntry_thenReturnSavedTx() {
 
-        final Tx tx = Tx.builder().ref(ref).date(LocalDate.now()).time(LocalTime.now()).descr("IFT Withdrawal").type("C").amount(89.99D).build();
+        final Tx tx = Tx.builder().ref(ref)
+                .date(LocalDate.now())
+                .time(LocalTime.now())
+                .descr("IFT Withdrawal").type("C")
+                .amount(BigDecimal.valueOf(89.99)).build();
 
         txRepo.save(tx);
         final Tx txResp = txRepo.findById(1L).get();
 
         assertNotNull(txResp);
         assertEquals("IFT Withdrawal", txResp.getDescr());
-        assertEquals(89.99, txResp.getAmount().doubleValue());
+        assertEquals(BigDecimal.valueOf(89.99), txResp.getAmount());
         assertEquals(ref, txResp.getRef());
     }
 }
